First pass
With a dummy example / copy of the modified thing. This is pretty
hardcoded but I just wanted to get a feel for libcst. The challenge will
be making this more modular and also could be tricky if I want to
implement any of this for lint rules that don't keep track of / point at
a specific object in the AST (but those generally might not require such
a thing, as they are often whitespace / formatting tyope changes).